Add check for root before remove_dir_all
- move info https://blog.rust-lang.org/2022/01/20/cve-2022-21658.html

+++ src/uninstall.rs
@@ -3,6 +3,7 @@ use super::package::get_local_dir;
 use super::spinner::spinner;
 use std::fs;
 use std::path::Path;
+use nix::unistd::Uid;
 
 pub fn uninstall(package: &str) -> std::io::Result<()> {
     let spin = spinner("Uninstalling...".to_string());
@@ -15,10 +16,14 @@ pub fn uninstall(package: &str) -> std::io::Result<()> {
         std::process::exit(0);
     }
 
-    fs::remove_dir_all(package_name)?;
-    debug!("Successfully removed package");
-
-    spin.finish_with_message("Done ✔");
+    // Check for root before running remove_dir_all
+    if !Uid::effective().is_root() {
+        fs::remove_dir_all(package_name)?;
+        debug!("Successfully removed package");
+        spin.finish_with_message("Done ✔");
+    } else {
+        spin.finish_with_message("Please do not run uninstall with root.")
+    }
 
     Ok(())
 }
